server/pyproject.toml
[tool.poetry]
name = "django-arcade-server"
version = "0.1.0"
description = ""
authors = ["Kurtis Fields <kurtisfields@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
ariadne-django = "^0.2.0"
loguru = "^0.5.3"
channels = "^3.0.4"
python-dateutil = "^2.8.2"
django-simple-bulma = "^2.4.0"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
